style: :art: formatting and minor stylistic edits (#3)
These are very minor, but they help to make the website look and work a
bit better.

- I made the first two columns of the program be smaller, and the other
two larger. Plus gave it a stripe to make it easier to view, but that's
entirely personal, you can remove it by removing the `.striped` from the
table caption section.
- Since it is a book format, it works best (I think because of it
generating the `sec-` labels) that the title of the page be a `# `
header rather than in the YAML with `title: `.